AuthenticationPrincipal 추가해서 시큐리티 인증, 인가 테스트 완료

+++ springboot-server/src/main/java/com/example/springbootserver/Todo/service/TodoService.java
@@ -1,15 +1,15 @@
 package com.example.springbootserver.todo.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.ObjectUtils;
 
 import com.example.springbootserver.todo.model.Todo;
 import com.example.springbootserver.todo.model.TodoRepository;
+import com.example.springbootserver.core.auth.session.MyUserDetails;
 import com.example.springbootserver.core.exception.Exception400;
+import com.example.springbootserver.core.exception.Exception403;
 import com.example.springbootserver.core.exception.Exception500;
 import com.example.springbootserver.todo.dto.TodoReq;
 import com.example.springbootserver.todo.dto.TodoReq.TodoSave;
@@ -28,41 +28,34 @@ public class TodoService {
     }
 
     @Transactional
-    public List<Todo> findByUserId(final Long id) {
+    public List<Todo> findByUserId(MyUserDetails userDetails) {
         try {
-            return todoRepository.findByUserId(id)
+            return todoRepository.findByUserId(userDetails.getUser().getId())
                 .orElseThrow(() -> new Exception400(null,"조회 데이터가 없습니다."));
         } catch (Exception500 e) {
             throw new Exception500("조회에 실패했습니다.");
         }
     }
 
-   @Transactional
-   public List<Todo> findAll() {
-       try {
-           List<Todo> result = todoRepository.findAll();
-
-           if (ObjectUtils.isEmpty(result)) {
-               result = new ArrayList<>();
-           }
-           return result;
-       } catch (Exception500 e) {
-           throw new Exception500("조회에 실패했습니다.");
-       }
-   }
-
     @Transactional
-    public Todo findbyId(final Long id) {
+    public Todo findbyId(final Long id, MyUserDetails userDetails) {
         try {
-            return todoRepository.findById(id)
+            Todo todo = todoRepository.findById(id)
                     .orElseThrow(() -> new Exception400(null,"조회 데이터가 없습니다."));
+            if ( todo.getUserId() != userDetails.getUser().getId()){
+                throw new Exception403("권한 필요");
+            }
+            return todo ;
         } catch (Exception500 e) {
             throw new Exception500("조회에 실패했습니다.");
         }
     }
 
     @Transactional
-    public Todo save(final TodoReq.TodoSave todoSave) {
+    public Todo save(final TodoReq.TodoSave todoSave, MyUserDetails userDetails) {
+        if ( todoSave.getUserId() != userDetails.getUser().getId()){
+            throw new Exception403("권한 필요");
+        }
         try {
             Todo todo = TodoSave.toEntity(todoSave);
             todoRepository.save(todo);
@@ -74,7 +67,10 @@ public class TodoService {
     }
 
     @Transactional
-    public Todo update(final TodoReq.TodoUpdate todoUpdate) {
+    public Todo update(final TodoReq.TodoUpdate todoUpdate, MyUserDetails userDetails) {
+        if ( todoUpdate.getUserId() != userDetails.getUser().getId()){
+            throw new Exception403("권한 필요");
+        }
         try {
             Todo todo = TodoReq.TodoUpdate.toEntity(todoUpdate);
             todoRepository.save(todo);
@@ -87,10 +83,13 @@ public class TodoService {
 
 
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, MyUserDetails userDetails) {
         try {
             Todo todo = todoRepository.findById(id)
                     .orElseThrow(() -> new Exception400(null,"조회 데이터가 없습니다."));
+            if ( todo.getUserId() != userDetails.getUser().getId()){
+                throw new Exception403("권한 필요");
+            }
             todoRepository.delete(todo);
             log.info("Todo delete 완료");
         } catch (Exception500 e) {
@@ -98,4 +97,19 @@ public class TodoService {
             throw new Exception500("삭제에 실패했습니다.");
         }
     }
+
+//    @Transactional
+//    public List<Todo> findAll() {
+//        try {
+//            List<Todo> result = todoRepository.findAll();
+
+//            if (ObjectUtils.isEmpty(result)) {
+//                result = new ArrayList<>();
+//            }
+//            return result;
+//        } catch (Exception500 e) {
+//            throw new Exception500("조회에 실패했습니다.");
+//        }
+//    }
+
 }
